Label: Hardcore 2024 for household income.
We don't want to use lastYear, because we need to wait for the tax, before asking a new year.
