+++ survey/locales/fr/end.yaml
@@ -5,7 +5,7 @@
 household.ownership: Est-ce que votre ménage est locataire ou propriétaire de votre
     logement?
 household.income: >-
-    Tranche de revenu avant impôts (brut) du ménage, en {{lastYear}}?<br /><span class="_pale
+    Tranche de revenu avant impôts (brut) du ménage, en 2024?<br /><span class="_pale
     _oblique">Facultatif</span>
 end.wouldLikeToParticipateInOtherSurveysChaireMobilite: >-
     Seriez-vous intéressés à <strong>participer à d'autres études</strong> menées
