+++ survey/references/questionnaire_list_fr.txt
@@ -395,7 +395,7 @@ Est-ce que votre ménage est locataire ou propriétaire de votre logement?
 - Autre
 - Je préfère ne pas répondre
 
-Tranche de revenu avant impôts (brut) du ménage, en {{lastYear}}?
+Tranche de revenu avant impôts (brut) du ménage, en 2024?
 Facultatif
 - Moins de 10 000 $
 - 10 000$ à 19 999$
